% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crunch_mean_by.R
\name{crunch_mean_by}
\alias{crunch_mean_by}
\title{Helper that gets the mean by some variables}
\usage{
crunch_mean_by(ds, mean_var, by_vars)
}
\arguments{
\item{ds}{a Dataset}

\item{mean_var}{string alias of a single variable you want to get a mean of}

\item{by_vars}{string aliases of variables to group by}
}
\value{
an object of class `CrunchCube`
}
\description{
Helper that gets the mean by some variables
}
